analysis/requirements.txt
pandas
sklearn
matplotlib
mysql-connector==2.1.4
